Настраивает наследование
